Move DDL generation logic from CLI to core library (jetdb)
Extract DDL generation (SQLite, PostgreSQL, MySQL, Access) from
jetdb-cli into jetdb core as a pub mod ddl, enabling reuse by
other consumers. The CLI ddl/ directory is replaced by a thin
wrapper module that delegates to the core implementation.
